requirements.txt
requests
python-dotenv
python-telegram-bot==20.0  # Specify version for compatibility
apscheduler
pytz